requirements.txt
# Core dependencies
pyaudio>=0.2.11
numpy>=1.24.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Whisper dependencies
openai-whisper>=20231117

# Optional: For better performance
torch>=2.0.0